Cleanup after merge.
