+++ libraries/lambda_handlers/client_linter_callback_handler.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals, print_function
+from libraries.client.client_linter_callback import ClientLinterCallback
 from libraries.lambda_handlers.handler import Handler
-from libraries.linters.client_linter_callback import ClientLinterCallback
 
 
 class ClientLinterCallbackHandler(Handler):
